+++ src/main/java/ru/vyarus/dropwizard/guice/test/jupiter/ext/GuiceyExtensionsSupport.java
@@ -13,7 +13,7 @@ import org.junit.platform.commons.util.ReflectionUtils;
 import ru.vyarus.dropwizard.guice.hook.ConfigurationHooksSupport;
 import ru.vyarus.dropwizard.guice.hook.GuiceyConfigurationHook;
 import ru.vyarus.dropwizard.guice.injector.lookup.InjectorLookup;
-import ru.vyarus.dropwizard.guice.test.jupiter.param.ClientSupport;
+import ru.vyarus.dropwizard.guice.test.ClientSupport;
 import ru.vyarus.dropwizard.guice.test.util.HooksUtil;
 
 import java.lang.reflect.Field;