+++ src/main/java/ru/vyarus/dropwizard/guice/test/spock/UseDropwizardApp.java
@@ -3,6 +3,7 @@ package ru.vyarus.dropwizard.guice.test.spock;
 import io.dropwizard.Application;
 import org.spockframework.runtime.extension.ExtensionAnnotation;
 import ru.vyarus.dropwizard.guice.hook.GuiceyConfigurationHook;
+import ru.vyarus.dropwizard.guice.test.ClientSupport;
 import ru.vyarus.dropwizard.guice.test.spock.ext.DropwizardAppExtension;
 
 import java.lang.annotation.ElementType;
@@ -11,16 +12,24 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Dropwizard app extension. Works almost the same as {@link io.dropwizard.testing.junit.DropwizardAppRule}, but
- * application instance is created for all tests in class (as if rule would be used with @ClassRule annotation)
- * <p>Services will be injected into the specification based on regular Guice annotations. {@code @Share} may
- * be used to define common injection points for all tests in class.</p>
- * <p>Note: {@code setupSpec()} fixture is called after application start and {@code cleanupSpec()} before
- * application tear down.</p>
- * <p>Extension behaviour is the same as spock-guice module.</p>
+ * Dropwizard spock extension. Starts dropwizard application before all tests in class and shutdown after them.
+ * <p>
+ * Gucie injections will work on test class (just annotate required fields with {@link javax.inject.Inject}.
+ * {@code @Share} may be used to define common injection points for all tests in class.
+ * <p>
+ * Note: {@code setupSpec()} fixture is called after application start and {@code cleanupSpec()} before
+ * application tear down.
+ * <p>
+ * Extension would also recognize static test fields (including super classes):
+ * <ul>
+ * <li>{@link GuiceyConfigurationHook} - hook from field will be registered</li>
+ * <li>{@link ru.vyarus.dropwizard.guice.test.ClientSupport} field will be injected with client instance. Note that
+ * only generic client may be used (to call 3rd party external services), as application's web part is not started.</li>
+ * </ul>
+ * <p>
+ * Internally based on {@link io.dropwizard.testing.DropwizardTestSupport}.
  *
  * @author Vyacheslav Rusakov
- * @see io.dropwizard.testing.junit.DropwizardAppRule for details
  * @since 03.01.2015
  */
 @Retention(RetentionPolicy.RUNTIME)
@@ -46,10 +55,45 @@ public @interface UseDropwizardApp {
     /**
      * Hooks provide access to guice builder allowing complete customization of application context
      * in tests.
+     * <p>
+     * Additional hooks could be declared in static test fields:
+     * {@code static GuiceyConfigurationHook HOOK = { it.disableExtensions(Something.class)}}.
      *
      * @return list of hooks to use
-     * @see UseGuiceyHooks to declare base hooks in base test class
      * @see GuiceyConfigurationHook for more info
      */
     Class<? extends GuiceyConfigurationHook>[] hooks() default {};
+
+    /**
+     * Enables random ports usage. Supports both simple and default dropwizard servers. Random ports would be
+     * set even if you specify exact configuration file with configured ports (option overrides configuration).
+     * <p>
+     * To get port numbers in test use {@link ClientSupport} static field:
+     * <pre>{@code
+     * static ClientSupport client
+     *
+     * static setupSpec() {
+     *     String baseUrl = "http://localhost:" + client.getPort();
+     *     String baseAdminUrl = "http://localhost:" + client.getAdminPort();
+     * }
+     * }</pre>
+     * Or use client target methods directly.
+     *
+     * @return true to use random ports
+     */
+    boolean randomPorts() default false;
+
+    /**
+     * Specifies rest mapping path. This is the same as specifying {@link #configOverride()}
+     * {@code "server.rootMapping=/something/*"}. Specified value would be prefixed with "/" and, if required
+     * "/*" applied at the end. So it would be correct to specify {@code restMapping = "api"} (actually set value
+     * would be "/api/*").
+     * <p>
+     * This option is only intended to simplify cases when custom configuration file is not yet used in tests
+     * (usually early PoC phase). It allows you to map servlet into application root in test (because rest is no
+     * more resides in test). When used with existing configuration file, this parameter will override file definition.
+     *
+     * @return rest mapping (empty string - do nothing)
+     */
+    String restMapping() default "";
 }