+++ src/main/java/ru/vyarus/dropwizard/guice/test/spock/ext/GuiceyConfigurationExtension.java
@@ -4,17 +4,16 @@ import org.spockframework.runtime.extension.AbstractAnnotationDrivenExtension;
 import org.spockframework.runtime.model.SpecInfo;
 import ru.vyarus.dropwizard.guice.hook.GuiceyConfigurationHook;
 import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyHooks;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
+import ru.vyarus.dropwizard.guice.test.util.HooksUtil;
 
 /**
  * {@link UseGuiceyHooks} extension implementation.
  *
  * @author Vyacheslav Rusakov
  * @since 12.04.2018
+ * @deprecated additional hooks may be declared in static test fields
  */
+@Deprecated
 public class GuiceyConfigurationExtension extends AbstractAnnotationDrivenExtension<UseGuiceyHooks> {
 
     private Class<? extends GuiceyConfigurationHook>[] confs;
@@ -27,22 +26,9 @@ public class GuiceyConfigurationExtension extends AbstractAnnotationDrivenExtens
     @Override
     public void visitSpec(final SpecInfo spec) {
         final GuiceyConfigurationHookInterceptor interceptor =
-                new GuiceyConfigurationHookInterceptor(instantiate(confs));
+                new GuiceyConfigurationHookInterceptor(HooksUtil.create(confs));
         // hook interceptor MUST go first, otherwise guicey will start without these changes
         spec.getTopSpec().getSharedInitializerInterceptors().add(0, interceptor);
         spec.getTopSpec().addCleanupSpecInterceptor(interceptor);
-    }
-
-    @SafeVarargs
-    public static final List<GuiceyConfigurationHook> instantiate(
-            final Class<? extends GuiceyConfigurationHook>... confs) {
-        return Arrays.stream(confs).map(conf -> {
-            try {
-                return conf.newInstance();
-            } catch (Exception e) {
-                throw new IllegalStateException("Failed to instantiate guicey hook: "
-                        + conf.getSimpleName(), e);
-            }
-        }).collect(Collectors.toList());
     }
 }