+++ src/test/groovy/ru/vyarus/dropwizard/guice/test/jupiter/dw/ClientSupportDwTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import ru.vyarus.dropwizard.guice.GuiceBundle;
 import ru.vyarus.dropwizard.guice.test.jupiter.TestDropwizardApp;
-import ru.vyarus.dropwizard.guice.test.jupiter.param.ClientSupport;
+import ru.vyarus.dropwizard.guice.test.ClientSupport;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;